pyproject.toml
[tool.poetry]
name = "admissions-agent"
version = "0.1.0"
description = "Telegram bot Admissions Agent ALT for the university admissions committee"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    { include = "src" },
]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
aiogram = "^3.5.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.30"}
pydantic = "^2.7.1"
google-genai = "^0.8.0"
chromadb = "^0.5.0"
python-dotenv = "^1.0.1"
aiosqlite = "^0.20.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.1"
pytest-asyncio = "^0.23.0"
ruff = "^0.4.4"
httpx = "^0.27.0"
black = "^24.4.2"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]  # Line too long

[tool.black]
line-length = 88
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 88
